Move common code into common.
